Creation/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.13
PyJWT==2.10.1
pytest==8.3.5
httpx==0.28.1
